- fix done mouse simulator - fix save related poistion by decimal

+++ OnymojiAuto/Code/Model/PointColor.cs
@@ -3,8 +3,8 @@
     public class PointColor
     {
         public string id { get; }
-        public int x { get; }
-        public int y { get; }
+        public decimal x { get; }
+        public decimal y { get; }
         public decimal color { get; }
 
         public PointColor(string id)
@@ -12,7 +12,7 @@
             this.id = id;
         }
 
-        public PointColor(string id, int x, int y, decimal color)
+        public PointColor(string id, decimal x, decimal y, decimal color)
         {
             this.id = id;
             this.x = x;
